+++ common/src/main/java/net/conczin/immersive_optimization/config/Config.java
@@ -28,6 +28,9 @@ public final class Config extends JsonConfig {
     public boolean enableEntities = true;
     public boolean enableBlockEntities = true;
 
+    // Slow down entities and block entities in force-loaded chunks.
+    public boolean optimizeForceLoadedChunks = false;
+
     // Distance culling reduces the tick rate of entities when not visible due to render distance.
     // (Singleplayer only)
     public boolean enableDistanceCulling = true;