+++ forge/src/main/java/net/conczin/immersive_optimization/ImmersiveOptimization.java
@@ -5,6 +5,7 @@ import net.minecraftforge.fml.common.Mod;
 @Mod(Constants.MOD_ID)
 public class ImmersiveOptimization {
     public ImmersiveOptimization() {
+        CommonClass.forcedChunkLookup = new ForgeForcedChunkLookup();
         CommonClass.init();
     }
 }